Migrate testing suite to Vitest
