+++ package.json
@@ -28,7 +28,6 @@
     "@rollup/plugin-typescript": "^11.1.6",
     "@types/expect": "^24.3.0",
     "@types/lodash": "^4.14.144",
-    "@types/mocha": "^10.0.0",
     "@types/node": "^18.7.14",
     "@typescript-eslint/eslint-plugin": "^7.1.1",
     "@typescript-eslint/parser": "^7.1.1",
@@ -39,11 +38,11 @@
     "is-url": "^1.2.4",
     "is-uuid": "^1.0.2",
     "lodash": "^4.17.15",
-    "mocha": "^10.0.0",
     "np": "^10.0.0",
     "prettier": "^3.2.5",
     "rollup": "^4.12.1",
-    "typescript": "^4.8.3"
+    "typescript": "^4.8.3",
+    "vitest": "^1.6.0"
   },
   "scripts": {
     "build": "rm -rf ./{dist} && rollup --config ./rollup.config.js",
@@ -55,9 +54,9 @@
     "lint:eslint": "eslint '{src,test}/*.{js,ts}'",
     "lint:prettier": "prettier --list-different '**/*.{js,json,ts}'",
     "release": "npm run build && npm run lint && np",
-    "test": "npm run build && npm run test:types && npm run test:mocha",
-    "test:mocha": "mocha --require ./test/register.cjs --require source-map-support/register ./test/index.ts",
+    "test": "npm run build && npm run test:types && npm run test:vitest",
     "test:types": "tsc --noEmit && tsc --project ./test/tsconfig.json --noEmit",
+    "test:vitest": "vitest run",
     "watch": "npm run build -- --watch"
   },
   "keywords": [
