+++ test/tsconfig.json
@@ -1,4 +1,7 @@
 {
   "extends": "../tsconfig.json",
-  "include": ["./**/*.ts"]
+  "include": ["./**/*.ts"],
+  "compilerOptions": {
+    "skipLibCheck": true
+  }
 }
